% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoU-classify-grid-rural.R
\name{classify_grid_rural}
\alias{classify_grid_rural}
\title{Create the DEGURBA grid cell classification of rural cells}
\usage{
classify_grid_rural(data, classification, value = 1)
}
\arguments{
\item{data}{path to the directory with the data, or named list with the data as returned by function \code{\link[=DoU_preprocess_grid]{DoU_preprocess_grid()}}}

\item{classification}{SpatRaster. A grid with the classification of urban centres and urban clusters to which the classification of rural cells will be added. Note that the grid will be adapted in-place.}

\item{value}{integer. Value assigned to rural cells in the resulting grid}
}
\value{
SpatRaster with the grid cell classification of rural cells
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{classify_grid_rural()} has been renamed to \code{DoU_classify_grid_rural()} to create a more consistent API and to better indicate that this function is specifically designed to classify rural cells in the context of the DEGURBA classification.
}
\keyword{internal}
