% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_zones_to_grid.R
\name{convert_zones_to_grid}
\alias{convert_zones_to_grid}
\title{Convert zones to a grid}
\usage{
convert_zones_to_grid(zones, referencegrid, id = NULL)
}
\arguments{
\item{zones}{character / object of class sf. Path to the vector layer with spatial zones, or an object of class sf with spatial zones}

\item{referencegrid}{Spatraster. Reference grid used to rasterize the values of the zones.}

\item{id}{character. The column that contains the values of the zones. If \code{NULL} the first column is used.}
}
\value{
SpatRaster with the zones
}
\description{
Convert the value of zones associated with a vector layer to a grid layer.
}
\examples{
# load data for urban proxies and zones 
proxies <- load_proxies_belgium()
zones <- flexurba::units_belgium

# convert Belgian provinces ('GID_2') to a zonal grid
gridded_zones <- convert_zones_to_grid(zones, proxies$pop, 'GID_2')
terra::plot(gridded_zones)
}
