% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoU-get-grid-parameters.R
\name{get_grid_parameters}
\alias{get_grid_parameters}
\title{Get the parameters for the DEGURBA grid cell classification}
\usage{
get_grid_parameters(level1 = TRUE)
}
\arguments{
\item{level1}{logical. Whether to return the standard parameters of level 1 of the Degree of Urbanisation (\code{TRUE}), or level 2 of the Degree of Urbanisation (\code{FALSE}).}
}
\value{
named list with the standard parameters
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{get_grid_parameters()} has been renamed to \code{DoU_get_grid_parameters()} to create a more consistent API and to better indicate that this function is specifically designed in the context of the DEGURBA classification with \code{classify_grid()}.
}
\keyword{internal}
