% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoU-load-grid-data-belgium.R
\name{load_grid_data_belgium}
\alias{load_grid_data_belgium}
\title{Load the grid data for Belgium to reconstruct DEGURBA classification}
\usage{
load_grid_data_belgium()
}
\value{
named list with the data of Belgium required for the grid classification of the Degree of Urbanisation.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{load_grid_data_belgium()} has been renamed to \code{DoU_load_grid_data_belgium()} to create a more consistent API and to better indicate that this function is specifically designed for reconstructing the DEGURBA classification with \code{DoU_classify_grid()}.
}
\keyword{internal}
