% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-threshold.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Identify urban boundaries by applying an relative threshold}
\usage{
apply_threshold(
  grid,
  type = "predetermined",
  threshold_value = NULL,
  fun = NULL,
  ...,
  zones = NULL,
  operator = "greater_than",
  smoothing = TRUE
)
}
\arguments{
\item{grid}{SpatRaster}

\item{type}{character. Either 'predefined' or 'data-driven'.}

\item{threshold_value}{numeric or vector. If \code{type="predefined"}, this threshold value is applied to derive urban boundaries. Be aware the threshold value should be in the same unit as the grid values. If \code{zones} is not \code{NULL}, you can also pass a vector of threshold values which should be applied in the respective zones.}

\item{fun}{character / function. If \code{type="data-driven"}, this function is used to derive the threshold value. Either as character: 'min', 'max', 'mean', 'median', or 'pX' where X is a percentile value (e.g., p25 for the 25\% percentile value"), or for relatively small grids a custom function.}

\item{...}{additional arguments passed to fun}

\item{zones}{character / SpatRaster / sf object. If  with the zones for which a separate threshold value will be derived.}

\item{operator}{character. Operator used to enforce the threshold. Either 'greater_than', 'greater_or_equal', 'smaller_than', 'smaller_or_equal' or 'equals'.}

\item{smoothing}{logical. Whether to smooth the edges of the boundaries. If \code{TRUE}, the the function \code{apply_majority_rule} will be employed to smooth the boundaries.}
}
\value{
named list with the following elements:
\itemize{
\item \code{rboundaries}: SpatRaster in which cells that are part of an urban unit have a value of '1'
\item \code{vboundaries}: sf object with the urban units as polygons
\item \code{threshold}: dataframe with per zone the threshold value that is applied
\item \code{zones}: SpatRaster with the zones
}
}
\description{
TODO
}
