% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_proxies_belgium.R
\name{load_proxies_belgium}
\alias{load_proxies_belgium}
\title{Load the data for three urban proxies for Belgium}
\usage{
load_proxies_belgium()
}
\value{
named list with gridded population, built-up area and night-time light data for Belgium.
}
\description{
The function loads a population and built-up grid from the \href{https://human-settlement.emergency.copernicus.eu/download.php}{GLobal Human Settlement Layer} and a night-time light grid from the \href{https://eogdata.mines.edu/products/vnl/#annual_v2}{Earth Observation Group}.

The specification of the source data is as follows:

The night-time light grid was processed with the following code to ensure that it aligns with the population and built-up layer.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{TODO
}\if{html}{\out{</div>}}
}
\examples{
load_proxies_belgium()
}
