% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-belgium.R
\docType{data}
\name{units_belgium}
\alias{units_belgium}
\title{GADM spatial units of Belgium}
\format{
\subsection{\code{units_belgium}}{

A object of class \code{sf} with the GADM spatial units for Belgium
}
}
\source{
\url{https://gadm.org/download_world.html}
}
\usage{
units_belgium
}
\description{
An object of class \code{sf} with the small spatial units of Belgium (downloaded from the \href{https://gadm.org/index.html}{GADM website}). The GADM data is filtered and converted to the Mollweide projection with the following code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# download the GADM data
download.file(
  "https://geodata.ucdavis.edu/gadm/gadm4.1/gadm_410-gpkg.zip",
  "data/GADM/gadm_410-gpkg.zip"
)

# unzip the file
unzip("data/GADM/gadm_410-gpkg.zip",
      files = "gadm_410.gpkg",
      exdir = "data/GADM")

# filter for Belgium, convert to Mollweide, and select only relevant attributes
units_belgium <-
  sf::st_read('data/GADM/gadm_410.gpkg', quiet = TRUE) \%>\%
  dplyr::filter(COUNTRY == "Belgium") \%>\%
  sf::st_transform('ESRI:54009') \%>\%
  dplyr::select(
    c(
      "UID",
      "GID_0",
      "NAME_0",
      "GID_1",
      "NAME_1",
      "GID_2",
      "NAME_2",
      "GID_3",
      "NAME_3",
      "GID_4",
      "NAME_4"
    )
  )
}\if{html}{\out{</div>}}
}
\keyword{datasets}
